+++ packages/db/tests/collection.test.ts
@@ -10,6 +10,11 @@ import {
   MissingUpdateHandlerError,
 } from "../src/errors"
 import { createTransaction } from "../src/transactions"
+import {
+  flushPromises,
+  mockSyncCollectionOptionsNoInitialState,
+  withExpectedRejection,
+} from "./utils"
 import type { ChangeMessage, MutationFn, PendingMutation } from "../src/types"
 
 describe(`Collection`, () => {
@@ -1197,5 +1202,68 @@ describe(`Collection`, () => {
     expect(collection.state.size).toBe(0)
     expect(collection.syncedData.size).toBe(0)
     expect(collection.syncedMetadata.size).toBe(0)
+  })
+
+  it(`open sync transaction isn't applied when optimistic mutation is resolved/rejected`, async () => {
+    type Row = { id: number; name: string }
+
+    const collection = createCollection(
+      mockSyncCollectionOptionsNoInitialState<Row>({
+        id: `repro-truncate-open-transaction`,
+        getKey: (r) => r.id,
+      })
+    )
+    const preloadPromise = collection.preload()
+
+    collection.utils.begin()
+    collection.utils.write({ type: `insert`, value: { id: 1, name: `one` } })
+    collection.utils.write({ type: `insert`, value: { id: 2, name: `two` } })
+    collection.utils.commit()
+    collection.utils.markReady()
+
+    await preloadPromise
+
+    expect(collection.state.size).toBe(2)
+
+    // start a transaction with a truncate, but don't commit it
+    collection.utils.begin()
+    collection.utils.truncate()
+
+    // expect the state to still be the same
+    expect(collection.state.size).toBe(2)
+
+    // we now do a local optimistic insert
+    collection.insert({ id: 3, name: `three` })
+
+    // we should immediately see the optimistic state
+    expect(collection.state.size).toBe(3)
+    expect(collection.state.get(3)?.name).toBe(`three`)
+
+    // we now reject the sync, this should trigger a rollback of the open transaction
+    // and the optimistic state should be removed
+    // it should *not* trigger the open sync transaction to be applied to the synced state
+    await withExpectedRejection(`trigger rollback`, () => {
+      collection.utils.rejectSync(new Error(`trigger rollback`))
+      return flushPromises()
+    })
+
+    // we should now be back to the original state, the optimistic state has
+    // been removed
+    expect(collection.state.size).toBe(2)
+
+    // write the new row back via sync
+    collection.utils.write({
+      type: `insert`,
+      value: { id: 3, name: `tree (from sync)` },
+    })
+
+    // expect the state to still be the same
+    expect(collection.state.size).toBe(2)
+
+    // now commit the sync transaction
+    collection.utils.commit()
+
+    // we truncated everything, so we should only have one item left that synced
+    expect(collection.state.size).toBe(1)
   })
 })